Create README.md
The objective was to send a secret text file encoded inside an image of bmp file format. Encoded the length of the secret text and then encoded the data into the LSB of the image bytes. The decoding process involves decoding the length and then decoding the text bit by bit. The final output is the secret text after decoding.